Move tests from context.rs to information_schema.rs
